Make sure errors are rethrown
We are catching certain errors in order to report them to Sentry with additional metadata, e.g. tags such as  for weather related errors. However, because these errors are being caught, they are no longer printed to the console (as uncaught exceptions usually are), which makes debugging extremely difficult—you don't know when errors are happpening, let alone where they are happening.

When catching errors for these purposes, we still want them to be handled the same as uncaught exceptions (i.e. logged to the console). This commit makes sure we re-throw the errors after they have been reported. We also make sure the error isn't reported twice (as uncaught exceptions are automatically reported by `window.onerror`/`unhandledRejection`

Esssentially this is a much better version of #9981. (Sorry about that, @piuccio and @rich-nguyen.)